Limit when the github action bazel builds run (#5071)
The bazel build are set up to run any time there is a push to any branch. This
cause the same workflow to be run twice when a PR is updated.

I'm changing the workflow to trigger on there is a push to "main", for the
continuous test, and leaving the pull_request trigger for the presubmit test.