build: allow getting version from VERSION file
When no git environment is available, the SPIRV-Tools version
cannot be computer correctly.
This was possible before with the CHANGES file.

The reason to move away from the CHANGES file was to have 1
source or truth (the git history), and prevent mistakes
as the PR to update the CHANGES file had to be done
on the commit we release.

Moving to a git-only method was fine, as long as users were
cloning the git repo. In the case of a github release, this
is bad, as the git history is not present in the zip files.

To solve the problem, the idea is to allow a VERSION file
to be present at the root of the project.
This file is by default **not** present. This file will be
added on the release branches, allowing the GitHub
releases to correctly report their version.

For users relying on SPIRV-Tools, but having no history,
the creation of a VERSION file (if not using the releases)
can be an alternative.