build: tool to generate a CHANGELOG from tags
This small utility parses the CHANGES file to generate
a CHANGELOG that can be used for GitHub releases.

This will be used by the GitHub workflow to make releases
automatic.

Given a CHANGES file path, a tag, the script extract the
editorialized changelog, and writes it to a given output file.