Remove duplicated code
After a conversation with Nathan, I noticed that the code in scalar replacement is
essentially replicating code in the type manager. I replace the code in scalar replacement with a call to the type manager.  I then apply the fix to the type manager.

I modified the meaning of `IsUniqueType` to match the way it is used. It
was used identify cases where there might be collisions in the
`type_to_id_` map.