val: allow DebugTypeComposite to forward reference
When an HLSL object has a member, the generated debug info
cannot be generated without any cycle:
 - the composite has a method as a member.
 - the method has the previously declared composite as scope.

This is not an issue as the spec clearly mentions members could
be a forward reference (which goes against the blanker no-forward
references statement).

This commit fixes the blanket check by verifying we are not in this
specific case.

Fixes #5229, Microsoft/DirectXShaderCompiler#5218

Signed-off-by: Nathan Gauër <brioche@google.com>