Fix 5191
Substitute variables for access chain function arguments. For each
function argument that is an access chain, create a new variable, copy
the access chain pointee into the variable before the function call,
substitute the variable in the function call, and copy the variable back
into the access chain pointee after the function call.

This is a workaround for NonSemantic.Shader.DebugInfo.100. DebugDeclare
expects the operand variable to be a result id of an OpVariable or
OpFunctionParameter. However, function arguments may contain
OpAccessChains during legalization which causes problems for
DebugDeclare.