Use windows 2019 to workaround bazel issue (#5261)
In the latest windows images, MSVC is not installed in the default
directory, and bazel is not able to find it.  See
https://github.com/actions/runner-images/issues/7675. Because of that, I
will change the github action so that is uses windows-2019 instead of
latest.

At the same time, I merged the build and test commands for the different
platforms because they are now the same.