linker: Add --use-first-entrypoint option
Currently spirv-link adds all entrypoints found in input files
into the output file. If multiple input files define the same
entrypoint, linking fails.

When this option is set, the entrypoints defined in the first
input file are used in the output and any entrypoints found in
subsequent input files are ignored.