Replace DEPS with git submodules
Git submodules are a more standard dependency management system than
DEPS files, and provide all of the features that SPIRV-Tools needs
without requiring depot_tools or custom scripts. The Chromium project is
also in the process of migrating from DEPS to git submodules.

I've left the custom sync and roll scripts in place as wrappers for the
equivalent git submodule commands to attempt to keep the user interface
as similar as possible.