Disable Wnonnull in BUILD.gn
There is an issue in the latest Android NDK where the `std::system` is
flagged as having a non-null parameter. This behaviour is currently used
by spirv-reduce and spirv-fuzz (and is documented as being allowed).

In order to allow building with the latest NDK disable the `Wnonnull`
warning in the BUILD.gn file.