[NFC] Refactor code to fold instruction in fold tests.
We repeat basicly the same code multiple times in the different types of
folding tests. This commit adds a function that builds the module, finds the
instruction to fold, and folds it. Doing the routine checks at the same
time. We also have a couple generatic functions for checking that an
instruction is a constant with the expected value.