Update .bazelversion
We need a newer version of bazel to get the latest `platforms` version, which includes the `fuchsia` platform, which the latest version of `googletest` uses. Will fix the tests for https://github.com/KhronosGroup/SPIRV-Tools/pull/5583.

If for any reason we can't update to the latest version of bazel, I'll submit a PR to override the default version of `platforms` by adding it to the `WORKSPACE`.