spirv-val: Disallow stores according to VUID 06924
Ensure that the validator rejects stores to objects of types
`OpTypeSampledImage` and `OpTypeAccelerationStructureKHR`, according
to `VUID-StandaloneSpirv-OpTypeImage-06924`.

Stores to objects of types `OpTypeImage` and `OpTypeSampler` should be
rejected too, but HLSL-to-SPIR-V translation by glslang and shaderc
seems to result in some violations; so leave validation of those as
future work.

Contributes to https://github.com/KhronosGroup/SPIRV-Tools/issues/4796

Signed-off-by: Sven van Haastregt <sven.vanhaastregt@arm.com>