spirv-as: Assume target from spvasm text
Assume the desired SPIR-V target version when the command line option is
not provided. This is especially important when the source being
assembled does not match the default (latest version), because it may
create an invalid SPIR-V binary result.

Also, clarify the wording for where a copy of spirv-headers is expected
(to match expected capitalization).