spirv-as: Assume target from spvasm text (#5893)
* spirv-as: Assume target from spvasm text

Assume the desired SPIR-V target version when the command line option is
not provided. This is especially important when the source being
assembled does not match the default (latest version), because it may
create an invalid SPIR-V binary result.

Also, clarify the wording for where a copy of spirv-headers is expected
(to match expected capitalization).

* fix formatting

* Add unit tests, static asserts on supported SPIR-V versions

Rework the algorithm so it accepts earlier termination of the version
string. That made testing easier and more intuitive.

* Fix compilation problems

---------

Co-authored-by: mmoult <30711895+mmoult@users.noreply.github.com>